requirements.txt
openmensa-api

selenium
webdriver-manager

python-telegram-bot  # pip install "python-telegram-bot[job-queue]"

langchain-core
ollama
langchain-ollama
langchain-groq
langchain-openai